--- LionHeart-iOS/LionHeart-iOS/Scenes/Article/ArticleDetail/Cells/ArticleDetailTableViewCell.swift
@@ -1,65 +1,0 @@
-//
-//  ArticleDetailTableViewCell.swift
-//  LionHeart-iOS
-//
-//  Created by uiskim on 2023/07/09.
-//  Copyright (c) 2023 ArticleDetail. All rights reserved.
-//
-
-import UIKit
-
-import SnapKit
-
-final class ArticleDetailTableViewCell: UITableViewCell, TableViewCellRegisterDequeueProtocol {
-    
-    var inputData: DummyModel? {
-        didSet {
-            /// action
-        }
-    }
-
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        // MARK: - 컴포넌트 설정
-        setUI()
-        
-        // MARK: - addsubView
-        setHierarchy()
-        
-        // MARK: - autolayout설정
-        setLayout()
-        
-        // MARK: - button의 addtarget설정
-        setAddTarget()
-        
-        // MARK: - delegate설정
-        setDelegate()
-    }
-    
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-private extension ArticleDetailTableViewCell {
-    func setUI() {
-        
-    }
-    
-    func setHierarchy() {
-        
-    }
-    
-    func setLayout() {
-        
-    }
-    
-    func setAddTarget() {
-        
-    }
-    
-    func setDelegate() {
-        
-    }
-}
